services/processing-engine/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.34.0
botocore==1.34.0
Pillow==10.2.0
python-magic==0.4.27
psycopg2-binary==2.9.9